Split dynamic Run handle
